level 4 5 and 6image
left side is cut off of 4-- this program doesn't work with water very
well. it's possible to take a super low res (1280x720) image of the
entire map or most of it, but it's low res and details are much harder
to see.

maps like 7, which have much too much water, will not work :(